Add all changes
