--- .env
@@ -1,7 +1,0 @@
-DB_USER=postgres
-DB_PASSWORD=mikeserver
-DB_HOST=localhost
-DB_PORT=5432
-DB_NAME=mydatabase
-
-JWT_SECRET=HJGDJEGFOEJKFE2GU
