--- .gitignore
@@ -1,5 +1,0 @@
-*.toml
-
-/tmp/main.exe
-
-.env
