+++ backend/.gitignore
@@ -1,0 +1,5 @@
+*.toml
+
+/tmp/main.exe
+
+.env
