+++ backend/middlewares/token
@@ -1,0 +1,29 @@
+package utils
+
+import (
+	"os"
+	"time"
+
+	"github.com/golang-jwt/jwt/v5"
+)
+
+// GenerateJWT creates a new JWT token
+func GenerateJWT(userID int, email string) (string, error) {
+	secret := os.Getenv("JWT_SECRET")
+	if secret == "" {
+		return "", nil
+	}
+
+	// Define claims
+	claims := jwt.MapClaims{
+		"id":    userID,
+		"email": email,
+		"exp":   time.Now().Add(time.Hour * 24).Unix(), // Token expires in 24 hours
+	}
+
+	// Create token
+	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
+
+	// Sign the token with the secret key
+	return token.SignedString([]byte(secret))
+}
